requirements.txt
Pillow==5.1.0
selenium==3.12.0
